tv-viewer/pyproject.toml
[tool.poetry]
name = "tv-viewer"
version = "0.1.0"
description = "Fast API wrappping tv-time"
authors = ["giraycoskun <giraycoskun@sabanciuniv.edu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.83.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
aioredis = "^2.0.1"
uvicorn = "^0.18.3"
logging = "^0.4.9.6"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"
beautifulsoup4 = "^4.11.1"
html5lib = "^1.1"
lxml = "^4.9.1"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
